add submodulesRemoteBranch option + submodules lfs
Add support for submodules remote branch option. This option allows to
check a branch with the same name on multiple submodules. The main
application is to send multiple pull-requests on submodules of the same
project while giving the CI system a way to build them altogether in
case there are incompatibilities.